roscam/requirements.txt
flask
pycapnp
tensorflow-gpu
keras
redis
numpy
scipy
pycapnp
requests
scikit-image
nltk
docopt
